refactor: move camera rig under board camera
